requirements.txt
fonttools
ufoLib2
ufo-extractor
ufo2ft
